盯播助手/stream-spider-python/requirements.txt
# 盯播助手Python爬虫和AI服务依赖

# Web框架
flask==2.3.3
flask-cors==4.0.0
flask-restx==1.1.0

# 异步框架
aiohttp==3.8.5
asyncio==3.4.3

# 数据抓取
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.11.2
playwright==1.36.0

# WebSocket
websockets==11.0.3
websocket-client==1.6.1

# 消息队列
pika==1.3.2
kafka-python==2.0.2

# 数据库
mysql-connector-python==8.1.0
redis==4.6.0
sqlalchemy==2.0.19
pymysql==1.1.0

# AI和NLP
openai==0.27.8
deepseek==0.1.0
transformers==4.31.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
sentence-transformers==2.2.2
jieba==0.42.1

# 配置管理
python-dotenv==1.0.0
pydantic==2.1.1

# 工具库
datetime==5.2
schedule==1.2.0
retry==0.9.2
loguru==0.7.0
python-dateutil==2.8.2
pytz==2023.3

# 监控和日志
prometheus-client==0.17.1
logging==0.5.1.2

# 测试
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# 其他
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
aiofiles==23.1.0
cryptography==41.0.2

# B站API
bilibili-api-python==16.2.0